// front-code/src/models/entity/class_grade_entity.ts
/*
 * ***************************************************************************************
 * author: XiaoLFeng(https://www.x-lf.com)
 * about:
 *   The project contains the source code of com.xlf.schedule.
 *   All source code for this project is licensed under the MIT open source license.
 * licenseStatement:
 *   Copyright (c) 2016-2024 XiaoLFeng. All rights reserved.
 *   For more information about the MIT license, please view the LICENSE file
 *     in the project root directory or visit:
 *   https://opensource.org/license/MIT
 * disclaimer:
 *   Since this project is in the model design stage, we are not responsible for any losses
 *     caused by using this project for commercial purposes.
 *   If you modify the code and redistribute it, you need to clearly indicate what changes
 *     you made in the corresponding file.
 *   If you want to modify it for commercial use, please contact me.
 * ***************************************************************************************
 */

import {ClassEntity} from "./class_entity.ts";

export type ClassGradeEntity = {
    class_grade_uuid: string;
    class_list: ClassEntity[];
    class_time_uuid: string;
    created_at: number;
    nickname: string;
    semester_begin: string;
    semester_end?: string;
    updated_at?: number;
    user_uuid: string;
}